Rename dashboad to dashboard
